+++ explorer/src/utils/auth/providers/index.ts
@@ -1,0 +1,4 @@
+import { Provider } from 'next-auth/providers'
+import { Discord } from './discord'
+
+export const providers: Provider[] = [Discord()]
